Code refactor, ChangeDisplayNameModal component created
