+++ apps/extension/src/screens/profile.screen.tsx
@@ -4,13 +4,9 @@ import {
   Avatar,
   Badge,
   Button,
-  Divider,
   FormControl,
-  FormErrorMessage,
   FormHelperText,
-  FormLabel,
   IconButton,
-  Input,
   Modal,
   ModalBody,
   ModalCloseButton,
@@ -30,8 +26,6 @@ import {
 import { useGetBasicUserInformation } from '../hooks/use-get-user-basic-information.hook';
 import React from 'react';
 import CustomInputBox from '../components/custom-input-box.component';
-import { BsPerson } from 'react-icons/bs';
-import { useChangeUserDisplayName } from '../hooks/use-change-user-displayname.hook';
 import { relativeTime } from '@wtm/utils';
 import { ArrowBackIcon } from '@chakra-ui/icons';
 import { useNavigation } from '../store';
@@ -45,10 +39,11 @@ import MaleTwo from '../assets/Avatars/male_two.png';
 import FemaleOne from '../assets/Avatars/female_one.png';
 import FemaleTwo from '../assets/Avatars/female_two.png';
 import ChangePasswordModal from '../components/change-password-modal.component';
+import ChangeDisplayNameModal from '../components/change-display-name-modal.component';
+
 export const ProfileScreen: React.FC<object> = () => {
   const { navigateBack } = useNavigation();
   const { basicUserInformationQuery } = useGetBasicUserInformation();
-  const { changeUserDisplayNameMutation } = useChangeUserDisplayName();
   const user = basicUserInformationQuery.data;
   const { changeUserAvatarMutation } = useChangeUserAvatar();
 
@@ -59,100 +54,6 @@ export const ProfileScreen: React.FC<object> = () => {
     FemaleTwo: FemaleTwo,
   };
 
-  const ChangeDisplayNameModal = () => {
-    const { isOpen, onOpen, onClose } = useDisclosure();
-    const [newDisplayName, setNewDisplayName] = useState('');
-    const isError =
-      newDisplayName === basicUserInformationQuery.data?.displayname;
-    const handleNewDisplaynameChange = (e: {
-      target: { value: React.SetStateAction<string> };
-    }) => setNewDisplayName(e.target.value);
-    useEffect(() => {
-      if (!isOpen) {
-        setNewDisplayName('');
-      }
-    }, [isOpen]);
-    const verifyFields = () => {
-      return (
-        newDisplayName &&
-        newDisplayName.length > 3 &&
-        newDisplayName != basicUserInformationQuery.data?.displayname
-      );
-    };
-    const handleSubmit = () => {
-      if (verifyFields()) {
-        changeUserDisplayNameMutation.mutate({
-          displayname: newDisplayName,
-        });
-      }
-    };
-
-    return (
-      <>
-        <Button
-          size='sm'
-          onClick={onOpen}
-          className='flex justify-between items-center gap-2'
-        >
-          <BsPerson /> Change display name
-        </Button>
-
-        <Modal
-          isOpen={isOpen}
-          onClose={onClose}
-          size={{ base: 'full', md: 'md' }}
-        >
-          <ModalOverlay />
-          <ModalContent>
-            <ModalHeader>Change the displayname</ModalHeader>
-            <ModalCloseButton />
-            <ModalBody>
-              <FormControl isInvalid={isError}>
-                <FormLabel>Actual Displayname</FormLabel>
-                <Input
-                  type='text'
-                  value={basicUserInformationQuery.data?.displayname}
-                  disabled
-                />
-                <Divider my={2} />
-                <FormLabel>New Displayname</FormLabel>
-                <Input
-                  type='text'
-                  value={newDisplayName}
-                  onChange={handleNewDisplaynameChange}
-                />
-                {!isError ? (
-                  <FormHelperText>
-                    Displayname must be different to actual displayname.
-                  </FormHelperText>
-                ) : (
-                  <FormErrorMessage>
-                    {'Displayname must be different to actual displayname.'}
-                  </FormErrorMessage>
-                )}
-              </FormControl>
-            </ModalBody>
-
-            <ModalFooter>
-              <div className='flex w-full justify-center md:justify-end'>
-                <Button
-                  colorScheme='blue'
-                  mr={3}
-                  onClick={handleSubmit}
-                  disabled={isError}
-                >
-                  Change
-                </Button>
-                <Button variant='ghost' onClick={onClose}>
-                  Close
-                </Button>
-              </div>
-            </ModalFooter>
-          </ModalContent>
-        </Modal>
-      </>
-    );
-  };
   const ChangeAvatarModal = () => {
     const images: ImageType[] = [
       {
